Move wallet connect to separate file
